requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
spacy>=3.7.0
scikit-learn>=1.4.0
google-genai>=1.0.0
pandas>=2.2.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
